docs: fix cancelrecurringinvoice version and counter types
The command does not exist at tag v25.09 (the plugin only gained
json_cancelrecurringinvoice in the v25.12 cycle), so the added
annotation is v25.12, not v25.09.  This also required updating
the recorded annotation in .msggen.json.

recurrence_counter and recurrence_start are parsed with
param_number (unsigned int) and are tu32 on the wire:

  plugins/fetchinvoice.c:867,873-874
    u32 *recurrence_counter, *recurrence_start;
    p_req("recurrence_counter", param_number, &recurrence_counter),
    p_opt("recurrence_start", param_number, &recurrence_start),
  wire/bolt12_wire.csv:93,95
    ...invreq_recurrence_counter,counter,tu32,
    ...invreq_recurrence_start,period_offset,tu32,

They were typed u64 and number; both are u32.

Changelog-None

+++ cln-rpc/src/model.rs
@@ -2452,9 +2452,9 @@ pub mod requests {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub payer_note: Option<String>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub recurrence_start: Option<f64>,
+	    pub recurrence_start: Option<u32>,
 	    pub offer: String,
-	    pub recurrence_counter: u64,
+	    pub recurrence_counter: u32,
 	    pub recurrence_label: String,
 	}
